[cc1 - cc2] Fix several errors in printf formats
We have to be careful that target dependant sizes must be printed
always with lld, because iin this way we can be sure it will be correct
always.

+++ cc1/code.c
@@ -258,7 +258,8 @@ emittype(Type *tp)
 		emitletter(tp);
 		putchar('\t');
 		emitletter(tp->type);
-		printf("\t#%c%d\n", sizettype->letter, tp->n.elem);
+		printf("\t#%c%lld\n",
+		       sizettype->letter, (long long) tp->n.elem);
 		return;
 	case PTR:
 		emittype(tp->type);
